build: switch back to *.flow.yaml from -flow.yaml
After working with it, `foo-bar.flow.yaml` is preferred and produces sensible
TypeScript modules like `foo-bar.flow.ts`, without having a stake in `-`
vs `_` for separation within names.
