+++ go/flow/.snapshots/TestLoadCollection
@@ -1,1 +1,1 @@
-(*flow.CollectionSpec)(collection:"testing/int-strings" schema_uri:"/examples/int-string-flow.yaml?ptr=/collections/testing~1int-strings/schema" key_ptrs:"/i" uuid_ptr:"/_meta/uuid" projections:<ptr:"/i" field:"eye" user_provided:true is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<field:"flow_document" inference:<types:"object" must_exist:true > > projections:<ptr:"/i" field:"i" is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<ptr:"/s" field:"s" inference:<types:"array" > > projections:<ptr:"/s/1" field:"s/1" inference:<types:"string" string:<> > > projections:<ptr:"/s/1" field:"sOne" user_provided:true inference:<types:"string" string:<> > > ack_json_template:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" )
+(*flow.CollectionSpec)(collection:"testing/int-strings" schema_uri:"/examples/int-string.flow.yaml?ptr=/collections/testing~1int-strings/schema" key_ptrs:"/i" uuid_ptr:"/_meta/uuid" projections:<ptr:"/i" field:"eye" user_provided:true is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<field:"flow_document" inference:<types:"object" must_exist:true > > projections:<ptr:"/i" field:"i" is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<ptr:"/s" field:"s" inference:<types:"array" > > projections:<ptr:"/s/1" field:"s/1" inference:<types:"string" string:<> > > projections:<ptr:"/s/1" field:"sOne" user_provided:true inference:<types:"string" string:<> > > ack_json_template:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" )
