+++ go/flow/.snapshots/TestLoadDerivedCollection
@@ -1,1 +1,1 @@
-(*flow.DerivationSpec)(collection:<collection:"testing/int-strings" schema_uri:"/examples/int-string-flow.yaml?ptr=/collections/testing~1int-strings/schema" key_ptrs:"/i" uuid_ptr:"/_meta/uuid" projections:<ptr:"/i" field:"eye" user_provided:true is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<field:"flow_document" inference:<types:"object" must_exist:true > > projections:<ptr:"/i" field:"i" is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<ptr:"/s" field:"s" inference:<types:"array" > > projections:<ptr:"/s/1" field:"s/1" inference:<types:"string" string:<> > > projections:<ptr:"/s/1" field:"sOne" user_provided:true inference:<types:"string" string:<> > > ack_json_template:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" > register_schema_uri:"/examples/int-string-flow.yaml?ptr=/collections/testing~1int-strings/derivation/register/schema" register_initial_json:"null" transforms:<derivation:"testing/int-strings" transform:"appendStrings" shuffle:<group_name:"derive/testing/int-strings/appendStrings" source_collection:"testing/int-string" source_partitions:<include:<labels:<name:"estuary.dev/collection" value:"testing/int-string" > > exclude:<> > source_uuid_ptr:"/_meta/uuid" shuffle_key_ptr:"/i" uses_source_key:true source_schema_uri:"/examples/int-string-flow.yaml?ptr=/collections/testing~1int-string/schema" uses_source_schema:true filter_r_clocks:true > publish_lambda:<typescript:"/derive/testing/int-strings/appendStrings/Publish" > > )
+(*flow.DerivationSpec)(collection:<collection:"testing/int-strings" schema_uri:"/examples/int-string.flow.yaml?ptr=/collections/testing~1int-strings/schema" key_ptrs:"/i" uuid_ptr:"/_meta/uuid" projections:<ptr:"/i" field:"eye" user_provided:true is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<field:"flow_document" inference:<types:"object" must_exist:true > > projections:<ptr:"/i" field:"i" is_primary_key:true inference:<types:"integer" must_exist:true > > projections:<ptr:"/s" field:"s" inference:<types:"array" > > projections:<ptr:"/s/1" field:"s/1" inference:<types:"string" string:<> > > projections:<ptr:"/s/1" field:"sOne" user_provided:true inference:<types:"string" string:<> > > ack_json_template:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" > register_schema_uri:"/examples/int-string.flow.yaml?ptr=/collections/testing~1int-strings/derivation/register/schema" register_initial_json:"null" transforms:<derivation:"testing/int-strings" transform:"appendStrings" shuffle:<group_name:"derive/testing/int-strings/appendStrings" source_collection:"testing/int-string" source_partitions:<include:<labels:<name:"estuary.dev/collection" value:"testing/int-string" > > exclude:<> > source_uuid_ptr:"/_meta/uuid" shuffle_key_ptr:"/i" uses_source_key:true source_schema_uri:"/examples/int-string.flow.yaml?ptr=/collections/testing~1int-string/schema" uses_source_schema:true filter_r_clocks:true > publish_lambda:<typescript:"/derive/testing/int-strings/appendStrings/Publish" > > )
